Move internalFormat to vscode-independent file
